Removing implicit integer warning

+++ test/src/test_object_xyz.cpp
@@ -23,16 +23,16 @@ void invoke_object(uint32_t max_symbols, uint32_t max_symbol_size, uint32_t mult
 
     typedef kodo::object_encoder<Encoder, Partitioning> object_encoder;
     typedef kodo::object_decoder<Decoder, Partitioning> object_decoder;
-    
+
     std::vector<char> data_in(max_symbols * max_symbol_size * multiplier);
     std::vector<char> data_out(max_symbols * max_symbol_size * multiplier, '\0');
 
     kodo::random_uniform<char> fill_data;
     fill_data.generate(&data_in[0], data_in.size());
-    
+
     typename Encoder::factory encoder_factory(max_symbols, max_symbol_size);
     typename Decoder::factory decoder_factory(max_symbols, max_symbol_size);
-    
+
     object_encoder obj_encoder(encoder_factory, kodo::storage(data_in));
     object_decoder obj_decoder(decoder_factory, obj_encoder.object_size());
 
@@ -40,7 +40,7 @@ void invoke_object(uint32_t max_symbols, uint32_t max_symbol_size, uint32_t mult
     EXPECT_EQ(multiplier, obj_decoder.decoders());
     EXPECT_TRUE(obj_encoder.encoders() == obj_decoder.decoders());
 
-    
+
     for(uint32_t i = 0; i < obj_encoder.encoders(); ++i)
     {
         typename Encoder::pointer encoder = obj_encoder.build(i);
@@ -53,11 +53,11 @@ void invoke_object(uint32_t max_symbols, uint32_t max_symbol_size, uint32_t mult
         // fully "filled" with data
         EXPECT_EQ(encoder->block_size(), encoder->bytes_used());
         EXPECT_EQ(decoder->block_size(), decoder->bytes_used());
-        
+
         EXPECT_TRUE(encoder->payload_size() == decoder->payload_size());
-        
+
         std::vector<uint8_t> payload(encoder->payload_size());
-        
+
         while(!decoder->is_complete())
         {
             encoder->encode( &payload[0] );
@@ -69,11 +69,11 @@ void invoke_object(uint32_t max_symbols, uint32_t max_symbol_size, uint32_t mult
             &data_out[0] + (i * encoder->block_size()), encoder->block_size());
 
         kodo::copy_symbols(storage, decoder);
-            
+
     }
-    
+
     EXPECT_TRUE(std::equal(data_in.begin(), data_in.end(), data_out.begin()));
-    
+
 }
 
 
@@ -90,19 +90,19 @@ void invoke_object_partial(uint32_t max_symbols,
 
     uint32_t object_size = max_symbols * max_symbol_size * multiplier;
     object_size -= (rand() % object_size);
-    
+
     typedef kodo::object_encoder<Encoder, Partitioning> object_encoder;
     typedef kodo::object_decoder<Decoder, Partitioning> object_decoder;
-    
+
     std::vector<char> data_in(object_size);
     std::vector<char> data_out(object_size, '\0');
 
     kodo::random_uniform<char> fill_data;
     fill_data.generate(&data_in[0], data_in.size());
-    
+
     typename Encoder::factory encoder_factory(max_symbols, max_symbol_size);
     typename Decoder::factory decoder_factory(max_symbols, max_symbol_size);
-    
+
     object_encoder obj_encoder(encoder_factory, kodo::storage(data_in));
     object_decoder obj_decoder(decoder_factory, obj_encoder.object_size());
 
@@ -111,7 +111,7 @@ void invoke_object_partial(uint32_t max_symbols,
     EXPECT_TRUE(obj_encoder.encoders() == obj_decoder.decoders());
 
     uint32_t bytes_used = 0;
-    
+
     for(uint32_t i = 0; i < obj_encoder.encoders(); ++i)
     {
         typename Encoder::pointer encoder = obj_encoder.build(i);
@@ -125,16 +125,16 @@ void invoke_object_partial(uint32_t max_symbols,
         EXPECT_TRUE(encoder->block_size() == decoder->block_size());
         EXPECT_TRUE(encoder->bytes_used() == decoder->bytes_used());
         EXPECT_TRUE(encoder->payload_size() == decoder->payload_size());
-        
+
         std::vector<uint8_t> payload(encoder->payload_size());
-        
+
         while(!decoder->is_complete())
         {
             encoder->encode( &payload[0] );
             decoder->decode( &payload[0] );
 
         }
-        
+
         kodo::mutable_storage storage = kodo::storage(
             &data_out[0] + bytes_used, decoder->bytes_used());
 
@@ -142,10 +142,10 @@ void invoke_object_partial(uint32_t max_symbols,
 
         bytes_used += decoder->bytes_used();
     }
-    
+
     EXPECT_EQ(bytes_used, object_size);
     EXPECT_TRUE(std::equal(data_in.begin(), data_in.end(), data_out.begin()));
-    
+
 }
 
 
@@ -161,7 +161,7 @@ void test_object_coders(uint32_t symbols, uint32_t symbol_size, uint32_t multipl
         kodo::full_rlnc8_encoder,
         kodo::full_rlnc8_decoder,
             kodo::rfc5052_partitioning_scheme>(symbols, symbol_size, multiplier);
-    
+
     invoke_object<
         kodo::full_rlnc16_encoder,
         kodo::full_rlnc16_decoder,
@@ -176,7 +176,7 @@ void test_object_coders(uint32_t symbols, uint32_t symbol_size, uint32_t multipl
         kodo::full_rlnc8_encoder,
         kodo::full_rlnc8_decoder,
             kodo::rfc5052_partitioning_scheme>(symbols, symbol_size, multiplier);
-    
+
     invoke_object_partial<
         kodo::full_rlnc16_encoder,
         kodo::full_rlnc16_decoder,
@@ -189,17 +189,17 @@ TEST(TestObjectCoder, construct_and_invoke_the_basic_api)
 {
     test_object_coders(32, 1600, 2);
     test_object_coders(1, 1600, 2);
-    
-    srand(time(0));
-    
+
+    srand(static_cast<uint32_t>(time(0)));
+
     uint32_t symbols = (rand() % 256) + 1;
     uint32_t symbol_size = ((rand() % 2000) + 1) * 2;
 
     // Multiplies the data to be encoded so that the object encoder
     // is expected to contain multiplier encoders.
     uint32_t multiplier = (rand() % 10) + 1;
-    
-    test_object_coders(symbols, symbol_size, multiplier);  
+
+    test_object_coders(symbols, symbol_size, multiplier);
 }
